Add pagination support to Cargo and optimize fetching
Implemented pagination support to better handle large volumes of Cargo, assisting in boosting the performance and scalability of the system. This was achieved via including a new method in ICargoRepository interface, implementing it in CargoRepository, and adding a new handler for fetching paged data.

Additional Tests were included to ensure the accurate performance of the new features. Also, minor adjustments were made in FetchByIdCommandHandler to convert it into FetchByIdQueryHandler for more precise terminology.